python-js/requirements.txt
boto3==1.41.2
botocore==1.41.2
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.15.0
six==1.17.0
urllib3==2.5.0
